Apply AppContainer permissions recursively also in subfolders
Fixes failing app startup with AppContainer sandboxing in Qt6 due to the "platforms\qwindows[d].dll" DLL not being detected during app startup. The reason for that is that the AppContainer process is only granted access to files and folders in the app folder, but not the files in subfolders like "platforms".